requirements.txt
click==6.6
configparser==3.5.0
decorator==4.0.10
funcsigs==1.0.2
future==0.16.0
futures==3.0.5
isodate==0.5.4
lxml==3.6.4
networkx==1.11
nibabel==2.1.0
nipype==0.13.1
numpy==1.13.0
pandas==0.20.1
prov==1.5.0
pydicom==0.9.9
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2017.2
rdflib==4.2.1
scipy==0.18.1
simplejson==3.10.0
six==1.10.0
traits==4.6.0
xvfbwrapper==0.2.8
bioread==1.0.4
